[PR-937] Add list of admins as env variable
Problem: Authorized users should be separated into two groups: editors and readers. User emails are not stored in the database.

Solution: Add admins' emails via an environment variable. After user authentication in the admin panel, his/her email is saved. Then it is compared with admins list to decide whether the user is an editor or a reader